requirements/tests.txt
-r production.txt

factory-boy
pytest-django
